js: migrate index page to vue component
We want to take more advantage of the components concept in a cleaner
way without using global variables to pass data.
