+++ app/assets/javascripts/modules/teams/index.js
@@ -1,2 +1,20 @@
-import './pages/index';
+import Vue from 'vue';
+
+import TeamsIndexPage from './pages/index';
+
 import './pages/show';
+
+$(() => {
+  if (!$('body[data-route="teams/index"]').length) {
+    return;
+  }
+
+  // eslint-disable-next-line no-new
+  new Vue({
+    el: '.vue-root',
+
+    components: {
+      TeamsIndexPage,
+    },
+  });
+});
