--- app/views/teams/partials/_panel.html.slim
@@ -1,23 +1,0 @@
-.panel.panel-default
-  .panel-heading
-    .row
-      .col-sm-6
-        h5 Teams you are member of
-      .col-sm-6.text-right
-        - if can_create_team?
-          <toggle-link text="Create new team" :state="state" state-key="newFormVisible" class="toggle-link-new-team"></toggle-link>
-
-  .panel-body
-    .table-responsive
-      <teams-table :teams="myTeams" :sortable="true" sort-by="name" teams-path="#{teams_path}"></teams-table>
-
-- if current_user.admin?
-  .panel.panel-default
-    .panel-heading
-      .row
-        .col-sm-6
-          h5 Other teams
-
-    .panel-body
-      .table-responsive
-        <teams-table :teams="otherTeams" :sortable="true" sort-by="name" teams-path="#{teams_path}"></teams-table>
